add metadata and active info to tests related to databases and models
