+++ api/tests/unit/datasets/test_like_dataset.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from api.src.usecases.datasets import toggle_like_dataset
 from api.src.errors import DatasetDoesNotExistError
-from api.src.models import User, Dataset
+from api.src.models import User, Dataset, Metadata
 
 
 @pytest.fixture
@@ -29,6 +29,14 @@ def dataset():
         source="http://test@m.com",
         license="test",
         files="123",
+        active=True,
+        metadata=Metadata(
+                description="test 3",
+                authors=["test"],
+                source="http://test@m",
+                license="test",
+                files="123",
+            ),
     )
